+++ app/models/FileOperation.ts
@@ -1,4 +1,5 @@
 import { computed } from "mobx";
+import { FileOperationFormat, FileOperationType } from "@shared/types";
 import { bytesToHumanReadable } from "@shared/utils/files";
 import BaseModel from "./BaseModel";
 import User from "./User";
@@ -16,7 +17,9 @@ class FileOperation extends BaseModel {
 
   size: number;
 
-  type: "import" | "export";
+  type: FileOperationType;
+
+  format: FileOperationFormat;
 
   user: User;
