+++ server/queues/tasks/CleanupExpiredFileOperationsTask.test.ts
@@ -1,9 +1,6 @@
 import { subDays } from "date-fns";
+import { FileOperationState, FileOperationType } from "@shared/types";
 import { FileOperation } from "@server/models";
-import {
-  FileOperationState,
-  FileOperationType,
-} from "@server/models/FileOperation";
 import { buildFileOperation } from "@server/test/factories";
 import { setupTestDatabase } from "@server/test/support";
 import CleanupExpiredFileOperationsTask from "./CleanupExpiredFileOperationsTask";
