+++ server/queues/tasks/CleanupExpiredFileOperationsTask.ts
@@ -1,8 +1,8 @@
 import { subDays } from "date-fns";
 import { Op } from "sequelize";
+import { FileOperationState } from "@shared/types";
 import Logger from "@server/logging/Logger";
 import { FileOperation } from "@server/models";
-import { FileOperationState } from "@server/models/FileOperation";
 import BaseTask, { TaskPriority } from "./BaseTask";
 
 type Props = {
